examples/pyproject.toml
[tool.poetry]
name = "example-dcm-schememo"
version = "0.1.0"
description = "An example project demonstrating the usage of dcm-schememo."
authors = ["oboenikui <oboenikuitwitter@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dcm-schememo = { path = ".." }
icalendar = "^5.0.7"
pywebview = "^5.4"
gkeepapi = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"